#2434: Fix search-context=initial (++)
